ginseg1.4
git-svn-id: svn+ssh://svn.creatis.insa-lyon.fr/svn/vip/Portal/trunk@4047 97043df2-9288-40f0-b43d-ea39296b91c7

+++ vip-query/src/main/java/fr/insalyon/creatis/vip/query/server/dao/business/QueryData.java
@@ -54,7 +54,7 @@ public class QueryData implements QueryDAO {
                 
                 while (rs2.next()) {
                Timestamp date=rs2.getTimestamp("dateCreation");
-               Long version=rs2.getLong("queryVersion");
+               Integer version=rs2.getInt("queryVersion");
                
                     queries.add(new String[]{rs.getString("queryName"),date.toString(),version.toString(),rs2.getString("queryversionID")});
                 }
